third_party/WebKit/LayoutTests/editing/deleting/forward-delete.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
  '<div contenteditable>|foo bar</div>',
  selection => {
    selection.document.execCommand('ForwardDelete');
    selection.document.execCommand('ForwardDelete');
    selection.document.execCommand('ForwardDelete');
    selection.document.execCommand('ForwardDelete');
  },
  '<div contenteditable>|bar</div>'),
  'Delete selected editable nodes that are not in both the start and end editable roots');
</script>
